Rename split.rs to global_affiliate_txs.rs
